+++ GadgetEngine/Graphics/Text/TextMesh.cpp
@@ -23,7 +23,7 @@ int TextMesh::GetTotalWidthInPixels() const{
 	int totalWidthInPixels = 0;
 	for(char c : text){
 		FreetypeFontCharacter ch = font->GetCharacters().at(c);
-		totalWidthInPixels += (ch.advanceX >> 6);
+		totalWidthInPixels += (static_cast<int>(ch.advanceX) >> 6);
 	}
 
 	return totalWidthInPixels;
@@ -35,7 +35,7 @@ int TextMesh::GetTotalHeightInPixels() const{
 		FreetypeFontCharacter ch = font->GetCharacters().at(c);
 
 		if(ch.rows > totalHeightInPixels){
-			totalHeightInPixels = ch.rows;
+			totalHeightInPixels = static_cast<int>(ch.rows);
 		}
 	}
